requirements.txt
# Core frameworks
fastapi>=0.110.0
uvicorn>=0.27.0
streamlit>=1.35.0

# Database
sqlmodel>=0.0.14
sqlite3-adapter>=0.0.1

# API utilities
python-multipart>=0.0.9
pydantic>=2.6.0
httpx>=0.27.0

# File processing
python-magic>=0.4.27
pdfplumber>=0.10.3
pandas>=2.2.0

# AI integration
openai>=1.12.0
aiosqlite>=0.19.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.5

# Security
python-jose>=3.3.0
python-dotenv>=1.0.0

# Utilities
tqdm>=4.66.2
loguru>=0.7.2